Reorganize `pants/backend/python`
[ci skip-rust]

[ci skip-build-wheels]
