+++ pants-plugins/internal_plugins/releases/register.py
@@ -3,7 +3,7 @@
 
 from packaging.version import Version
 
-from pants.backend.python.rules.run_setup_py import SetupKwargs, SetupKwargsRequest
+from pants.backend.python.goals.setup_py import SetupKwargs, SetupKwargsRequest
 from pants.engine.fs import DigestContents, GlobMatchErrorBehavior, PathGlobs
 from pants.engine.rules import Get, collect_rules, rule
 from pants.engine.target import Target
