--- src/python/pants/backend/python/README.md
@@ -1,1 +1,0 @@
-[Pants documentation](https://www.pantsbuild.org/docs/python-backend) has moved...
