+++ src/python/pants/backend/python/goals/BUILD
@@ -9,7 +9,6 @@ python_tests(
   timeout=180,
 )
 
-
 python_integration_tests(
   name='integration',
   uses_pants_run=True,
