+++ src/python/pants/backend/python/goals/setup_py_test.py
@@ -6,9 +6,7 @@ from typing import Iterable, Type
 
 import pytest
 
-from pants.backend.python.python_artifact import PythonArtifact
-from pants.backend.python.rules import python_sources
-from pants.backend.python.rules.run_setup_py import (
+from pants.backend.python.goals.setup_py import (
     AmbiguousOwnerError,
     DependencyOwner,
     ExportedTarget,
@@ -24,20 +22,25 @@ from pants.backend.python.rules.run_setup_py import (
     SetupPyChrootRequest,
     SetupPySources,
     SetupPySourcesRequest,
+    declares_pkg_resources_namespace_package,
     determine_setup_kwargs,
+    distutils_repr,
     generate_chroot,
     get_exporting_owner,
     get_owned_dependencies,
     get_requirements,
     get_sources,
+    is_python2,
     validate_args,
 )
+from pants.backend.python.macros.python_artifact import PythonArtifact
 from pants.backend.python.target_types import (
     PythonBinary,
     PythonDistribution,
     PythonLibrary,
     PythonRequirementLibrary,
 )
+from pants.backend.python.util_rules import python_sources
 from pants.core.target_types import Files, Resources
 from pants.engine.addresses import Address
 from pants.engine.fs import Snapshot
@@ -743,3 +746,94 @@ def test_validate_args() -> None:
 
     validate_args(("sdist",))
     validate_args(("bdist_wheel", "--foo"))
+
+
+def test_distutils_repr() -> None:
+    testdata = {
+        "foo": "bar",
+        "baz": {"qux": [123, 456], "quux": ("abc", b"xyz"), "corge": {1, 2, 3}},
+        "various_strings": ["x'y", "aaa\nbbb"],
+    }
+    expected = """
+{
+    'foo': 'bar',
+    'baz': {
+        'qux': [
+            123,
+            456,
+        ],
+        'quux': (
+            'abc',
+            'xyz',
+        ),
+        'corge': {
+            1,
+            2,
+            3,
+        },
+    },
+    'various_strings': [
+        'x\\\'y',
+        \"\"\"aaa\nbbb\"\"\",
+    ],
+}
+""".strip()
+    assert expected == distutils_repr(testdata)
+
+
+@pytest.mark.parametrize(
+    "python_src",
+    [
+        "__import__('pkg_resources').declare_namespace(__name__)",
+        "\n__import__('pkg_resources').declare_namespace(__name__)  # type: ignore[attr-defined]",
+        "import pkg_resources; pkg_resources.declare_namespace(__name__)",
+        "from pkg_resources import declare_namespace; declare_namespace(__name__)",
+    ],
+)
+def test_declares_pkg_resources_namespace_package(python_src: str) -> None:
+    assert declares_pkg_resources_namespace_package(python_src)
+
+
+@pytest.mark.parametrize(
+    "python_src",
+    [
+        "",
+        "import os\n\nos.getcwd()",
+        "__path__ = 'foo'",
+        "import pkg_resources",
+        "add(1, 2); foo(__name__); self.shoot(__name__)",
+        "declare_namespace(bonk)",
+        "just nonsense, not even parseable",
+    ],
+)
+def test_does_not_declare_pkg_resources_namespace_package(python_src: str) -> None:
+    assert not declares_pkg_resources_namespace_package(python_src)
+
+
+@pytest.mark.parametrize(
+    "constraints",
+    [
+        ["CPython>=2.7,<3"],
+        ["CPython>=2.7,<3", "CPython>=3.6"],
+        ["CPython>=2.7.13"],
+        ["CPython>=2.7.13,<2.7.16"],
+        ["CPython>=2.7.13,!=2.7.16"],
+        ["PyPy>=2.7,<3"],
+    ],
+)
+def test_is_python2(constraints):
+    assert is_python2(constraints)
+
+
+@pytest.mark.parametrize(
+    "constraints",
+    [
+        ["CPython>=3.6"],
+        ["CPython>=3.7"],
+        ["CPython>=3.6", "CPython>=3.8"],
+        ["CPython!=2.7.*"],
+        ["PyPy>=3.6"],
+    ],
+)
+def test_is_not_python2(constraints):
+    assert not is_python2(constraints)
