+++ src/python/pants/backend/python/macros/BUILD
@@ -1,0 +1,6 @@
+# Copyright 2020 Pants project contributors (see CONTRIBUTORS.md).
+# Licensed under the Apache License, Version 2.0 (see LICENSE).
+
+python_library()
+
+python_tests(name="tests")
